Implement Warning the user after successful added if the series was 100 before to reach the maximum 10 digits SeriesNumber in OR

+++ Repository/ReceiptRepo.cs
@@ -32,7 +32,7 @@ namespace Accounting_System.Repository
             }
         }
 
-        public async Task<int> GetLastSeriesNumberOR()
+        public async Task<long> GetLastSeriesNumberOR()
         {
             var lastInvoice = await _dbContext
                 .OfficialReceipts
